+++ lib/event_bus_riverpod.dart
@@ -1,8 +1,11 @@
 export 'src/bus_metadata.dart';
 export 'src/event_bus_action.dart';
+export 'src/event_bus_action_for_global.dart';
 export 'src/event_bus_extension.dart';
+export 'src/event_bus_global.dart';
 export 'src/event_bus_identifier.dart';
 export 'src/event_bus_provider.dart';
+
 export 'src/listener_disposable.dart';
 export 'src/sub_event_action.dart';
 export 'src/sub_event_identifier.dart';