+++ lib/src/event_bus_action_for_global.dart
@@ -1,0 +1,26 @@
+import 'package:event_bus_riverpod/src/event_bus_action.dart';
+import 'package:event_bus_riverpod/src/event_bus_provider.dart';
+import 'package:event_bus_riverpod/src/event_bus_singleton.dart';
+import 'package:event_bus_riverpod/src/sub_event_action.dart';
+
+/// [EventBusAction] implementation that uses the global
+/// [EventBusSingleton] instead of a Riverpod [Ref].
+class EventBusActionForGlobal<T> extends EventBusAction<T>
+    with EventBusActionMixin<T> {
+  EventBusActionForGlobal({required super.event});
+
+  @override
+  EventBusCore get eventBus => EventBusSingleton.getInstance().core;
+}
+
+/// [SubEventAction] implementation that uses the global
+/// [EventBusSingleton] instead of a Riverpod [Ref].
+class SubEventActionForGlobal<T> extends SubEventAction<T>
+    with SubEventActionMixin<T> {
+  SubEventActionForGlobal({required super.identifier}) {
+    eventBus.setHistorySize(identifier.key, identifier.historySize);
+  }
+
+  @override
+  EventBusCore get eventBus => EventBusSingleton.getInstance().core;
+}